extract shared types from contextmenu
